refactor(mobile): split media viewer controls
